Split GuildUpdateService into separate services with proper error logging
