+++ src/Commands/Events/ErrorLoggingPostExecutionEvent.cs
@@ -1,7 +1,6 @@
 using JetBrains.Annotations;
 using Microsoft.Extensions.Logging;
 using Remora.Discord.Commands.Contexts;
-using Remora.Discord.Commands.Extensions;
 using Remora.Discord.Commands.Services;
 using Remora.Results;
 
@@ -31,14 +30,7 @@ public class ErrorLoggingPostExecutionEvent : IPostExecutionEvent
     public Task<Result> AfterExecutionAsync(
         ICommandContext context, IResult commandResult, CancellationToken ct = default)
     {
-        if (!commandResult.IsSuccess && !commandResult.Error.IsUserOrEnvironmentError())
-        {
-            _logger.LogWarning("Error in slash command execution.\n{ErrorMessage}", commandResult.Error.Message);
-            if (commandResult.Error is ExceptionError exerr)
-            {
-                _logger.LogError(exerr.Exception, "An exception has been thrown");
-            }
-        }
+        _logger.LogResult(commandResult, "Error in slash command execution.");
 
         return Task.FromResult(Result.FromSuccess());
     }
